Auto-format everything, checkpoint work on Previews
I got it building with some of the work I'm doing with Previews, but the work isn't right yet. I want to be able to see a preview when the user declines to give up permissions. I haven't figured out yet why that isn't working, but I do have a bit of code around changing the Preview UI I want to save.

Also, pre-commit is formatting everything which is cool.

I should have split this commit, but whatever.