Add save button to new note page
This took some really crazy detective work. The issue is that I was
using the same view for both the "edit note" and "add note" flow, but I
was getting to each view differently, one via tabs and the other via a
navigation link from a list of notes. This makes the platform do
different things with the toolbar. Also, it does different things with
the dismiss() call - specifically it doesn't dismiss without a
navigation link.

Instead I split them apart and let each one be specialized, including
having a callback that programmatically changes the currently active
tab, which simulates the navigation dismiss.

In doing this I also added a "toast-like" UI element, though it's hard
to see now, I think.