Switch to sqlite backing store.
It's much faster than SwiftData, I have much tighter control over it,
and I've already figured out how to send updates from the background
actor to the frontend, which was the part I was most scared about.

I'm currently limiting the front end to just 10 items to keep it snappy.