Bind map bounds to list of notes
This is pretty messy, as I was exploring a bunch of ways to do this.
This is just the first that worked. I now show only the list of notes
that correspond to the map boundaries and add items to the map markers
only if they fit in the map. Both are fast now.