Simplify relationship between url background downloader and manager
Previously we were analyzing the URL of the request to try to determine
where the state machine was at. Now instead we use continuations and
keep track of the tasks which allows the download manager to be written
in a straightforward async style.

This is all to facilitate surfacing sync information.