Show background network progress in main view
I removed ModelActor as it added nothing but an extra layer of actor.

I had to create a system for moving the temporary URLs to permanent URLs
before parsing since the OS removes the temporary file as soon as my
urlSession callback exits, even if I try to resume a continuation from
it. This led to "no such file or directory" errors when I tried to parse
the file.

I've added an enum for different statuses that are shown in the UI.