Remove nested navigationview for note editor
Okay, so I was staring too much at the layout and reading stuff online
and realized that I probably had added the navigationview to overcome
what I was doing with the navigation view at the root of the app.

This demonstrates that I'm correct, though it doesn't yet help me solve
all the stuff that it breaks around determining the correct menu to
display.