Move note buffer and toasts into the main model
This is because the save button is now made available by the root view.
This means we need to move the data to somewhere the root view can get
it - the root model.

Overall this will be a better pattern, I think, for various reasons, but
it requires a pretty big rework.

This is just one step in several.