Don't include the non-existent migrations bundle
This is meant to make it possible to bundle SQL scripts together and use
them for migrations. We don't do our migrations that way, prefering
swift code instead. Since the bundle is an empty directory we can't add
it to git, so this throws exceptions at runtime on build machines that
don't manually create the bundle file.

That's bad.