Restore picture saving, remove image table
This is a big change, I'm removing the old picture table with its
reference to the note table and creating a new table to take its place
to store picture data with location information.

Because this removes the entire image table I'm also removing and
renaming any functions related to the table.