Make BreadcrumbModel an @Observable class
I'm not totally certain on the reason for this, but tests have shown
that the camera UI instability is related to accessing the properties of
this struct. There are no errors, no exceptions, no log messages, it's
just that if I both have a callback for location that modifies the
properties of this class and a view that reads those properties, even
for a log message, the camera's control widgets disappear. This is
becasue the AVCaptureSession switches its isRunning from true to false.

It's been expensive to debug.