Precalculate hex summaries, render via MKMapView
This is a huge rework of how we do overlay hexes. Specifically we are
now pre-calculating the summary data and storing it in the database for
fast recall. We the manually manage rendering of the hexes via an
MKMapView which allows for a large number of optimizations over the
naive implementation of overlays we previously had.

Hat tip to https://fatbobman.com/en/posts/uikitinswiftui/ for a bunch of
the technical detail