Add basic knowledge graphing
There's a bunch of thrashing around at the project level. I was trying
to get to where unit tests would work, but I never sorted it out and
actually created a separate project so I could get unit test.

XCode. I hate XCode.

The logic here is pretty bad, it's just designed to accomplish the
requirements of a demo video, not actually fit-to-purpose.