Redo overview icon rendering system
This adds in a bunch of enums which allows me to defer deciding on the
exact rendering style to the code that controls the UI. That makes it
possible to try different multi-color scheme without trying to encode it
in a string.

I've translated the existing icons into this new system and also added
the new icons we eventually want to represent. I also make the icon
locations static rather than packing them for easier comparison.