Use preview data for photo note 'icon'
This does two things - first it passes the data of the photo preview as
captured by the AV stack and passes it through the Photo object allowing
us to write the file to disk.

Second, it attempts to use that file, but when not present, it generates
a thumbnail from the image it has.