Handle partial completions for sentence endings.
Adding the custom language model changed the way that the system gives
me partials. Instead of continuing to get a stream of words if I pause
for a few seconds I get an entirely new list of words as if the previous
ones weren't uttered. I therefore have to keep track of them and join
them together.