Break utterances into sentences, re-add explicit breeding flag
This gets us to test 4 successful, and also fixes a number of previous
test cases. I break the utterances into sentences because the lexer
can't figure out the sentence breaks if the next sentence is not a
complete sentence. This makes sense for a general-purpose lexer, but it
doesn't work for us due to the prevalence of terse technical utterances
that aren't actual sentences.